Fixed console warnings.
- Failed prop type: Invalid prop `temp[0]` of type `object` supplied to `Home`, expected an array.
- Failed prop type: Invalid prop `co2[0]` of type `object` supplied to `Home`, expected an array.
- The pseudo class ":first-child" is potentially unsafe when doing server-side rendering. Try changing it to ":first-of-type".

I believe the type declaration in `index.js` was just wrong -
`Scatter` expects `data` to be an array of `{x, y}` objects, and
that's what `static.config.js` generates, so it's what the type
declaration should expect.

I don't know what "potentially unsafe" means, but adopting the
recommendation to `:first-of-type` maintains the expected styling
and doesn't seem to do any harm.